requirements.txt
pandas
pyarrow
numpy
python-dateutil
requests
tqdm
duckdb
pyyaml
reportlab
